test: rename test-utils.tsx to testUtils.tsx for ls-lint
client/ .tsx files must be camelCase per .ls-lint.yml.
