+++ frontend/app/chat/page.test.tsx
@@ -93,7 +93,7 @@ vi.mock("@/components/ControlPanel", () => ({
   ),
 }));
 
-import { makeTextMessage } from "@/client/test-utils";
+import { makeTextMessage } from "@/client/testUtils";
 import { AssistantModel, AssistantTemperature } from "@/client/types/assistant";
 import Home from "./page";
